feat: Read for Purchase, Update Category, Update IPurchaseInterface

+++ HomeAccounting.Application/Commands/Categories/Handlers/Commands/DeleteCategoryCommandHandler.cs
@@ -9,20 +9,31 @@ namespace HomeAccounting.Application.Commands.Categories.Handlers.Commands
     public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryCommand, BaseServicesResponse>
     {
         private readonly ICategoryRepository _categoryRepository;
+        private readonly IPurchaseRepository _purchaseRepository;
 
-        public DeleteCategoryCommandHandler(ICategoryRepository categoryRepository)
+        public DeleteCategoryCommandHandler(ICategoryRepository categoryRepository, IPurchaseRepository purchaseRepository)
         {
             _categoryRepository = categoryRepository;
+            _purchaseRepository = purchaseRepository;
         }
 
         public async Task<BaseServicesResponse> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
         {
             var response = new BaseServicesResponse();
-            // var category = new Category(); заглушка для проверки.
-            // Когда сделаю CRUD для остальных, сюда запихать логику проверки, если категория есть в ПОКУПКАХ, её удалить нельзя.
             try
             {
                 var category = await _categoryRepository.GetById(request.Id);
+                var isCategoryUsed = await _purchaseRepository.IsCategoryUsedAsync(category.Id);
+
+                if (isCategoryUsed)
+                {
+                    response.Success = false;
+                    response.Id = category.Id;
+                    response.Message = "You cannot delete a category because it is being used";
+
+                    return response;
+                }
+
                 await _categoryRepository.Delete(category);
 
                 response.Id = category.Id;
